renamed server to service, created StdioServer
