+++ ruffd-core/src/lib.rs
@@ -3,9 +3,10 @@ extern crate lazy_static;
 
 mod notifications;
 mod requests;
-mod server;
+pub mod server;
 mod server_ops;
+mod service;
 
 pub const PKG_NAME: &str = env!("CARGO_PKG_NAME");
 pub const PKG_VERSION: &str = env!("CARGO_PKG_VERSION");
-pub use server::Server;
+pub use service::Service;
