+++ src/main.rs
@@ -1,8 +1,8 @@
-use ruffd_core::Server;
+use ruffd_core::server::StdioServer;
 use ruffd_types::tokio;
 
 #[tokio::main]
 async fn main() {
-    let mut server = Server::new();
-    server.run_stdio().await;
+    let mut server = StdioServer::default();
+    server.get_service_mut().run().await;
 }
